requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
joblib
kagglehub
streamlit
tensorflow
pillow
numpy
